requirement.txt
django==2.1.5
django-rest_framework
djangorestframework-jwt